Finished save information and account creation tests

+++ src/test/java/tests/VerifyUsersCanDeleteShippingInformation.java
@@ -1,6 +1,7 @@
 package tests;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.Cookie;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.sikuli.script.FindFailed;
@@ -13,7 +14,10 @@ import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 import patterns.DriverManager;
 import patterns.pageobject.addressBookPage;
+import utilities.Utils;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 public class VerifyUsersCanDeleteShippingInformation {
@@ -24,26 +28,32 @@ public class VerifyUsersCanDeleteShippingInformation {
 
     @BeforeTest
     public void beforeTest() throws FindFailed{
-        driver = DriverManager.getDriver(DriverManager.BrowserType.EDGE); // replace with your desired browser
+        driver = DriverManager.getDriver(DriverManager.BrowserType.CHROME); // replace with your desired browser
         //Login as admin to unlock functionalities
-        driver.get("https://demo.opencart.com/admin/");
-        WebElement inputName = driver.findElement(By.id("input-username"));
-        inputName.sendKeys("demo");
-        WebElement inputPassword = driver.findElement(By.id("input-password"));
-        inputPassword.sendKeys("demo");
-        WebElement login = driver.findElement(By.cssSelector(".btn"));
-        login.click();
+        Date today = new Date();
+        SimpleDateFormat format = new SimpleDateFormat("yyyyMMddHHmmss");
+        String formatedDate = format.format(today);
+        driver.get("https://demo.opencart.com/");
+        driver.manage().addCookie(new Cookie("OCSESSID","11c0f931cf"+formatedDate+"ec"));
+        driver.manage().addCookie(new Cookie("_ga","GA1.1.2123778129.1713796835"));
+        driver.manage().addCookie(new Cookie("_ga_X8G0BRFSDF","GS1.1.1713796835.1.0.1713796835.0.0.0"));
+        driver.manage().addCookie(new Cookie("_gcl_au","1.1.534898992.1713796834"));
+        driver.manage().addCookie(new Cookie("_gid","GA1.2.438931849.1713796835"));
+        driver.manage().addCookie(new Cookie("cf_clearance","zJ9wxfXGd6JiMI3czkXFs4.kzRi6IqvPGPR1BaphLjM-1713852454-1.0.1.1-XKiVE5CVgEaZJ6pwxaPFZvAbzObkzBLWVzgfCCZoPHgWbHPgp6V.HROlod2Rr0jRzg2O5vNoDLVqbRP0JC8Gnw"));
+        driver.manage().addCookie(new Cookie("currency","USD"));
     }
 
 
     @Test
     @Parameters({"email","password"})
-    public void UsersCanSaveShippingInformation(String email, String password) throws FindFailed {
+    public void UsersCanDeleteShippingInformation(String email, String password) throws FindFailed {
         final int MAX_ATTEMPTS = 20;
         boolean passedLogin = false;
+        boolean passedDelete = false;
+        int numberOfAddresses = 0;
+        int tryDelete = 0;
         for (int attempt = 0; attempt < MAX_ATTEMPTS; attempt++) {
             try {
-                driver.get("https://demo.opencart.com/");
                 addressBookPage addressBook = new addressBookPage(driver);
                 addressBook.clickNavMyAccountDropdown();
                 if(passedLogin){
@@ -51,14 +61,23 @@ public class VerifyUsersCanDeleteShippingInformation {
                 }else{
                     addressBook.clickNavLogin();
                     passedLogin = addressBook.login(email,password);
+                    Utils.takeSnapShot(driver, "src/resources/SaveShipping_and_BillingInformationResults/DeleteShippingInformationResults/1-loginSuccessful.png");
                 }
-                addressBook.clickAddressBookLink();
-                int numberOfAddresses = addressBook.getAllAddressesLength().size();
-                System.out.println(numberOfAddresses);
-                addressBook.deleteAddress();
-                addressBook.clickAddressBookLink();
+                if(passedDelete){
+                    addressBook.clickAddressBookLink();
+                }else{
+                    addressBook.clickAddressBookLink();
+                    Utils.takeSnapShot(driver, "src/resources/SaveShipping_and_BillingInformationResults/DeleteShippingInformationResults/2-seeAddressBook.png");
+                    numberOfAddresses = addressBook.getAllAddressesLength().size();
+                    while(!passedDelete && tryDelete<5){
+                        passedDelete = addressBook.deleteAddress();
+                        tryDelete = tryDelete + 1;
+                    }
+                    Utils.takeSnapShot(driver, "src/resources/SaveShipping_and_BillingInformationResults/DeleteShippingInformationResults/3-deleteAddress.png");
+                    addressBook.clickAddressBookLink();
+                }
                 int newNumberOfAddresses = addressBook.getAllAddressesLength().size();
-                System.out.println(newNumberOfAddresses);
+                Utils.takeSnapShot(driver, "src/resources/SaveShipping_and_BillingInformationResults/DeleteShippingInformationResults/4-deletedAddress.png");
                 Assert.assertEquals(newNumberOfAddresses,(numberOfAddresses-1));
                 break;  // if successful, break the loop
             } catch (Exception e) {
